Move Java code to jvmMain directory
