+++ kotlinpoet/build.gradle.kts
@@ -31,7 +31,9 @@ spotless {
 }
 
 kotlin {
-  jvm()
+  jvm {
+    withJava()
+  }
 
   sourceSets {
     val commonMain by getting {
