+++ kotlinpoet/src/commonTest/kotlin/com/squareup/kotlinpoet/AnnotationSpecTest.kt
@@ -345,32 +345,6 @@ class AnnotationSpecTest {
     )
   }
 
-  @Test fun getOnValueArrayTypeMirrorShouldNameValueArg() {
-    val myClazz = compilation.elements
-      .getTypeElement(JavaClassWithArrayValueAnnotation::class.java.canonicalName)
-    val classBuilder = TypeSpec.classBuilder("Result")
-
-    myClazz.annotationMirrors.map { AnnotationSpec.get(it) }
-      .forEach {
-        classBuilder.addAnnotation(it)
-      }
-
-    assertThat(toString(classBuilder.build())).isEqualTo(
-      """
-            |package com.squareup.tacos
-            |
-            |import com.squareup.kotlinpoet.JavaClassWithArrayValueAnnotation
-            |import java.lang.Boolean
-            |import java.lang.Object
-            |
-            |@JavaClassWithArrayValueAnnotation.AnnotationWithArrayValue(value = arrayOf(Object::class,
-            |        Boolean::class))
-            |public class Result
-            |
-      """.trimMargin(),
-    )
-  }
-
   @Test fun getOnVarargMirrorShouldNameValueArg() {
     val myClazz = compilation.elements
       .getTypeElement(KotlinClassWithVarargAnnotation::class.java.canonicalName)
@@ -394,28 +368,6 @@ class AnnotationSpecTest {
         |import kotlin.Boolean
         |
         |@AnnotationSpecTest.AnnotationWithArrayValue(value = arrayOf(Object::class, Boolean::class))
-        |public class Result
-      """.trimMargin(),
-    )
-  }
-
-  @Test fun getOnValueArrayTypeAnnotationShouldNameValueArg() {
-    val annotation = JavaClassWithArrayValueAnnotation::class.java.getAnnotation(
-      JavaClassWithArrayValueAnnotation.AnnotationWithArrayValue::class.java,
-    )
-    val classBuilder = TypeSpec.classBuilder("Result")
-      .addAnnotation(AnnotationSpec.get(annotation))
-
-    assertThat(toString(classBuilder.build()).trim()).isEqualTo(
-      """
-        |package com.squareup.tacos
-        |
-        |import com.squareup.kotlinpoet.JavaClassWithArrayValueAnnotation
-        |import java.lang.Boolean
-        |import java.lang.Object
-        |
-        |@JavaClassWithArrayValueAnnotation.AnnotationWithArrayValue(value = arrayOf(Object::class,
-        |        Boolean::class))
         |public class Result
       """.trimMargin(),
     )
